🚚 move: 이벤트 개설 페이지 상수 폴더 이동
