+++ src/constants/openEvent/openEventConstants.ts
